requirements.txt
appnope==0.1.0
backcall==0.2.0
decorator==4.4.2
ipdb==0.13.3
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
numpy==1.19.2
pandas==1.1.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
psycopg2-binary==2.8.6
ptyprocess==0.6.0
Pygments==2.7.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
traitlets==5.0.4
wcwidth==0.2.5
